feat(engine): Functioning board (#3)
* Add `Square`

Add a `Square` class with color, rank, file, and (optionally) a piece
that it is occupied by.

* Add `Board`, `InvalidCoordinateException`, `CoordinateValidator`

* Refactor to use `Coordinate` record instead of `rank` and `file`

Also change `InvalidCoordinateException` to extend `IllegalArgumentException`

* Add `Move` record to contain information about a move

* Add `getRankFrom` method to `Board`; unify test imports and `getSaure`

This commit adds the `getRankFrom` method to `Board`, which returns the
squares in the rank in both directions.

It also unifies test imports (importing `static` methods from
`Assertions`), and also unifies the paramater order when dealing with
coordinates (`file` then `rank` — like in a chess coordinate).

* Add `getFileFrom` method to `Board`

* Add `getDiagonalsFrom` method to `Board`

* Add `getLsFrom` method to `Board`

* Change `Coordinate` to a class

This allows for more fine-grained control over `equals` and `toString`,
and also allows for simpler handling of `rank`: Data is saved at `byte`
but handled externally as an `int`.

* Refactor so all coordinate logic uses `Coordinate`

* Add `makeMove` to `Board`; `MoveResult`; `InvalidMoveException`

Also refactor `Move` to use `Coordinate` instead of `Square`.